created some cpuutility
